correcciones de 1era entrega realizadas

+++ src/managers/productsManagers.js
@@ -29,15 +29,13 @@ class ProductManager {
             const products = await this.getProducts()
             let resultado = products.find(producto => Number(producto.id) == Number(productId));
 
-            if (resultado) {
-                return resultado;
-            } else {
-                console.log('Producto no encontrado');
-                return null;
+            if (!resultado) {
+                throw new Error('El producto que buscas no existe');
             }
+            return resultado;
         }
         catch (error) {
-            return console.log('Producto no encontrado', error);
+            return console.log('Producto no encontrado.', error);
         }
     }
 
@@ -47,11 +45,6 @@ class ProductManager {
         try {
             let products = await this.getProducts()
             let newId = products.length > 0 ? products[products.length - 1].id + 1 : 1;
-            const exists = products.some(product => product.title === title);
-
-            if (exists) {
-                return console.log(`El producto con el nombre ${title} ya existe`);
-            }
 
             const product = {
                 title: title,
@@ -61,16 +54,22 @@ class ProductManager {
                 thumbnail: thumbnail,
                 stock: stock,
                 id: newId,
-                code: newId + title,
+                code: 'SKU' + newId,
                 status: true
             }
+
+            const exists = products.some(p => p.code === product.code);
+            if (exists) {
+                throw new Error(`El producto con el código ${code} ya existe`)
+            }
+
             products.push(product)
             await fs.promises.writeFile(this.path, JSON.stringify(products), 'utf-8') // agrego mis nuevos productos en un json
             this.products = products
             console.log('Producto agregado correctamente');
             return product;
         } catch (error) {
-            console.log('Error al agregar el producto:', error);
+            console.log('Error al agregar producto:', error);
         }
     };
 
@@ -79,7 +78,6 @@ class ProductManager {
 
     // actualizar producto por id
     async updateProductById(productId, newDetails) {
-
         try {
             let products = await this.getProducts();
             let product = products.find(producto => producto.id === productId);
@@ -88,18 +86,31 @@ class ProductManager {
                 throw new Error('Producto no encontrado');
             }
 
-            products = products.map
-                (
-                    producto => producto.id === productId ? { ...producto, ...newDetails } : producto
-                );
+            // Verificar si se está intentando modificar el campo id o agregar nuevos campos
+            const camposRequeridos = ['title', 'category', 'description', 'price', 'thumbnail', 'stock'];
+            const newDetailsKeys = Object.keys(newDetails);
+
+            // Verificar si hay campos no permitidos o si el campo 'id' está presente
+            const camposInvalidos = newDetailsKeys.some(key => !camposRequeridos.includes(key) || key === 'id');
+            if (camposInvalidos) {
+                throw new Error('No se puede modificar el campo id ni agregar campos no permitidos');
+            }
+
+            // Actualizar el producto sin modificar el campo id
+            products = products.map(producto =>
+                producto.id === productId ? { ...producto, ...newDetails } : producto
+            );
+
             await fs.promises.writeFile(this.path, JSON.stringify(products), 'utf-8');
-            this.products = products
+            this.products = products;
             return this.products;
         } catch (error) {
             console.log('Error al actualizar producto', error);
             throw error;
         }
     }
+
+
 
     // borrar producto por su id
     async deleteProduct(productId) {
@@ -108,14 +119,14 @@ class ProductManager {
             let listaProductos = await this.getProducts()
             let resultado = listaProductos.filter(product => product.id !== productId);
 
-            if (resultado.length < listaProductos.length) {
+            if (resultado && (resultado.length < listaProductos.length)) {
                 await fs.promises.writeFile(this.path, JSON.stringify(resultado), 'utf-8');
                 console.log('Producto eliminado exitosamente');
                 this.products = resultado
                 return this.products;
             }
             else {
-                console.log('Producto no encontrado');
+                return console.log('Producto no encontrado');
             }
 
         } catch (error) {
